grpc-postgres-customer/requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
asyncpg==0.29.0
